pyproject.toml
[project]
name = "python-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index>=0.12.11",
    "textual-plotext>=0.2.1",
    "ckanapi>=4.8",
    "ipykernel>=6.29.5",
    "numpy>=2.2.0",
    "scipy>=1.15.1",
    "pandas>=2.2.3",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.0.1"
]
